+++ lib/intl.dart
@@ -171,11 +171,11 @@ class Intl {
           String name,
           List args,
           String meaning}) =>
-      _message(message_str, locale, name, args);
+      _message(message_str, locale, name, args, meaning);
 
   /// Omit the compile-time only parameters so dart2js can see to drop them.
-  static _message(String message_str, String locale, String name, List args) {
-    return messageLookup.lookupMessage(message_str, locale, name, args);
+  static _message(String message_str, String locale, String name, List args, String meaning) {
+    return messageLookup.lookupMessage(message_str, locale, name, args, meaning);
   }
 
   /// Return the locale for this instance. If none was set, the locale will